+++ Parrafo.h
@@ -1,18 +1,22 @@
 #ifndef __PARRAFO_H__
 #define __PARRAFO_H__
+/**
+ * Esta clase almacena la información de un párrafo, en particular el offset de donde
+ * inicia y donde termina en el archivo correspondiente al documento.
+ */
 class Parrafo {
 private:
     int id;
-    int inicio;
-    int final;
+    long inicio;
+    long final;
 public:
-    Parrafo(int id, int inicio, int final);
+    Parrafo(int id, long inicio, long final);
     Parrafo(const Parrafo& otro);
     Parrafo() = default;;
     ~Parrafo() = default;
 
     int getId();
-    int getInicio();
-    int getFinal();
+    long getInicio();
+    long getFinal();
 };
 #endif //__PARRAFO_H__
